requirements.txt
requests
py_essentials
simplejson
pyyaml
coloredlogs
docker